feat: optimistic comments + route loading skeleton (motion engine A4)
- Comments: a posted comment shows instantly (optimistic), rendered with the SAME
  renderCommentMarkdown the server uses (no drift), overlaid via tested
  lib/comment-tree.ts mergeOptimisticComments, then reconciled by the authoritative
  refetch (which clears the overlay). Failed POST removes it + shows the error.
  Composer + sign-in buttons split into CommentForm.tsx (keeps Comments.tsx < 400).
- (blog)/loading.tsx: themed skeleton for instant nav feedback; pulse follows the
  motion engine (.skeleton, gated by data-motion + reduced-motion), sized to match
  a post (no shift on swap).
- Tests: comment-tree.test.ts (append top-level / nested reply / no-mutate).
