+++ src/app/globals.css
@@ -587,3 +587,20 @@ button:not(:disabled):active,
     transform: none;
   }
 }
+
+/* Loading skeleton block (route loading.tsx). Faint themed surface; the subtle
+   pulse follows the motion engine (gated by data-motion + reduced-motion), so it
+   degrades to a calm static placeholder. */
+.skeleton {
+  background: var(--c-rule);
+  border-radius: 0.375rem;
+}
+@media (prefers-reduced-motion: no-preference) {
+  :root[data-motion='on'] .skeleton {
+    animation: skeleton-pulse 1.4s var(--ease) infinite;
+  }
+}
+@keyframes skeleton-pulse {
+  0%, 100% { opacity: 0.55; }
+  50% { opacity: 1; }
+}
