+++ src/components/blog/Comments.tsx
@@ -1,30 +1,26 @@
 'use client'
 
 // Public comment island. The post page stays ISR/static; this fetches the comment
-// tree from /api/comments with `no-store`, so a just-posted comment shows on the
-// next load with NO cache in the way. After a successful post we simply refetch —
-// authoritative + instant, no optimistic reconciliation to get wrong.
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import { signIn, signOut } from 'next-auth/react'
+// tree from /api/comments with `no-store`. A just-posted comment shows INSTANTLY
+// (optimistic) using the SAME `renderCommentMarkdown` the server uses — so there is
+// no content drift — then the authoritative refetch replaces it and clears the
+// pending overlay. On failure the optimistic comment is removed + an error shown.
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { signOut } from 'next-auth/react'
 import type { PublicComment, SiteLang } from '@/types'
 import { t, formatDate } from '@/lib/i18n'
-import { Turnstile } from './Turnstile'
+import { renderCommentMarkdown } from '@/lib/comment-md'
+import { mergeOptimisticComments } from '@/lib/comment-tree'
+import { CommentForm, SignInButton, MAX, type Draft } from './CommentForm'
 
 type Viewer = { name: string } | null
 
-const MAX = 1000
 const MAX_DEPTH = 2
-
-type Draft = { name: string; email: string; website: string; content: string }
-const EMPTY: Draft = { name: '', email: '', website: '', content: '' }
 
 // Count real (non-tombstone) comments across the whole tree, for the heading.
 function countVisible(nodes: PublicComment[]): number {
   return nodes.reduce((n, c) => n + (c.deleted ? 0 : 1) + countVisible(c.replies), 0)
 }
-
-const inputClass =
-  't-small w-full rounded-lg border border-rule bg-bg px-3 py-2 text-text outline-none focus:border-heading'
 
 export function Comments({
   postSlug,
@@ -43,10 +39,12 @@ export function Comments({
 }) {
   const s = t(lang)
   const [comments, setComments] = useState<PublicComment[]>([])
+  const [pending, setPending] = useState<PublicComment[]>([])
   const [loaded, setLoaded] = useState(false)
   const [replyTo, setReplyTo] = useState<number | null>(null)
   const [viewer, setViewer] = useState<Viewer>(null)
   const oauthOn = googleAuth || facebookAuth
+  const tempId = useRef(-1) // optimistic ids count down so they never clash with real ones
 
   // Fetch the tree (no setState) so callers decide when to commit it to state —
   // keeps setState out of the synchronous effect body (deferred, after await).
@@ -95,25 +93,48 @@ export function Comments({
     }
   }, [oauthOn])
 
-  // POST a comment; on success refetch the tree and close any reply form.
+  // POST a comment. Show it optimistically (same renderer as the server), then
+  // reconcile with an authoritative refetch and drop the pending overlay. On any
+  // failure, remove the optimistic comment so the form can report the error.
   const submit = useCallback(
     async (parentId: number | null, draft: Draft, token: string | null): Promise<boolean> => {
-      const res = await fetch('/api/comments', {
+      const id = tempId.current--
+      const optimistic: PublicComment = {
+        id,
+        parentId,
+        name: viewer?.name || draft.name.trim(),
+        website: draft.website.trim() || undefined,
+        provider: 'manual',
+        contentHtml: renderCommentMarkdown(draft.content.slice(0, MAX)),
+        createdAt: new Date().toISOString(),
+        deleted: false,
+        replies: [],
+      }
+      setPending((p) => [...p, optimistic])
+      const ok = await fetch('/api/comments', {
         method: 'POST',
         headers: { 'content-type': 'application/json' },
         body: JSON.stringify({ postSlug, parentId, ...draft, turnstileToken: token }),
-      }).catch(() => null)
-      if (!res || !res.ok) return false
-      const json = await res.json().catch(() => null)
-      if (!json?.success) return false
-      setComments(await fetchTree())
+      })
+        .then((res) => res.ok && res.json())
+        .then((json) => !!json?.success)
+        .catch(() => false)
+      if (!ok) {
+        setPending((p) => p.filter((c) => c.id !== id))
+        return false
+      }
+      const fresh = await fetchTree()
+      setComments(fresh)
+      setPending((p) => p.filter((c) => c.id !== id))
       setReplyTo(null)
       return true
     },
-    [postSlug, fetchTree],
+    [postSlug, fetchTree, viewer],
   )
 
-  const count = useMemo(() => countVisible(comments), [comments])
+  // Optimistic comments overlaid on the authoritative tree, for render + count.
+  const displayed = useMemo(() => mergeOptimisticComments(comments, pending), [comments, pending])
+  const count = useMemo(() => countVisible(displayed), [displayed])
 
   return (
     <section className="mt-6">
@@ -147,11 +168,11 @@ export function Comments({
         )}
       </div>
 
-      {loaded && comments.length === 0 ? (
+      {loaded && displayed.length === 0 ? (
         <p className="mt-6 t-small text-meta">{s.commentsEmpty}</p>
       ) : (
         <ul className="mt-8 space-y-6">
-          {comments.map((c) => (
+          {displayed.map((c) => (
             <CommentNode
               key={c.id}
               comment={c}
@@ -262,123 +283,5 @@ function CommentNode({
         </ul>
       )}
     </li>
-  )
-}
-
-function SignInButton({ label, provider }: { label: string; provider: 'google' | 'facebook' }) {
-  return (
-    <button
-      type="button"
-      onClick={() => signIn(provider, { callbackUrl: window.location.href })}
-      className="t-small rounded-lg border border-rule px-3 py-2 text-text hover:bg-rule"
-    >
-      {label}
-    </button>
-  )
-}
-
-function CommentForm({
-  lang,
-  onSubmit,
-  turnstile,
-  turnstileSiteKey,
-  viewer = false,
-  autoFocus,
-}: {
-  lang: SiteLang
-  onSubmit: (draft: Draft, token: string | null) => Promise<boolean>
-  turnstile: boolean
-  turnstileSiteKey: string
-  viewer?: boolean
-  autoFocus?: boolean
-}) {
-  const s = t(lang)
-  const [draft, setDraft] = useState<Draft>(EMPTY)
-  const [busy, setBusy] = useState(false)
-  const [error, setError] = useState(false)
-  const [token, setToken] = useState<string | null>(null)
-  const set = (k: keyof Draft, v: string) => setDraft((d) => ({ ...d, [k]: v }))
-  const onToken = useCallback((tk: string | null) => setToken(tk), [])
-
-  // A signed-in viewer needs no identity fields and no Turnstile (the server
-  // trusts the session). Otherwise: name + email, and Turnstile (if on) must be
-  // solved before the content step — identity first so the gate is meaningful.
-  const showWidget = turnstile && !!turnstileSiteKey && !viewer
-  const identityOk = viewer || (!!draft.name.trim() && /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(draft.email.trim()))
-  const gateOpen = !showWidget || !!token
-  const valid = identityOk && draft.content.trim() && gateOpen
-
-  async function handle(e: React.FormEvent) {
-    e.preventDefault()
-    if (!valid || busy) return
-    setBusy(true)
-    setError(false)
-    const okied = await onSubmit({ ...draft, content: draft.content.slice(0, MAX) }, token)
-    setBusy(false)
-    if (okied) {
-      setDraft(EMPTY)
-      setToken(null) // tokens are single-use — force a fresh solve for the next one
-    } else {
-      setError(true)
-    }
-  }
-
-  return (
-    <form onSubmit={handle} className="space-y-2.5">
-      {!viewer && (
-        <div className="grid gap-2.5 sm:grid-cols-3">
-          <input
-            className={inputClass}
-            placeholder={s.commentName}
-            value={draft.name}
-            onChange={(e) => set('name', e.target.value)}
-            autoFocus={autoFocus}
-            maxLength={80}
-          />
-          <input
-            className={inputClass}
-            type="email"
-            placeholder={`${s.commentEmail} · ${s.commentEmailNote}`}
-            value={draft.email}
-            onChange={(e) => set('email', e.target.value)}
-            maxLength={120}
-          />
-          <input
-            className={inputClass}
-            placeholder={s.commentWebsite}
-            value={draft.website}
-            onChange={(e) => set('website', e.target.value)}
-          />
-        </div>
-      )}
-
-      {/* Gate: with Turnstile on, the widget appears once identity is filled and
-          must be solved before the comment box is shown. */}
-      {showWidget && !gateOpen ? (
-        identityOk && <Turnstile siteKey={turnstileSiteKey} onToken={onToken} />
-      ) : (
-        <>
-          <textarea
-            className={`${inputClass} min-h-[6rem] resize-y`}
-            placeholder={s.commentBody}
-            value={draft.content}
-            onChange={(e) => set('content', e.target.value.slice(0, MAX))}
-          />
-          <div className="flex items-center justify-between gap-3">
-            <span className="t-small text-meta">
-              {s.commentFormatHint} · {draft.content.length}/{MAX}
-            </span>
-            <button
-              type="submit"
-              disabled={!valid || busy}
-              className="t-small rounded-lg border border-heading bg-heading px-4 py-2 font-medium text-bg transition-opacity disabled:opacity-40"
-            >
-              {busy ? s.commentSubmitting : s.commentSubmit}
-            </button>
-          </div>
-        </>
-      )}
-      {error && <p className="t-small text-meta">{s.commentError}</p>}
-    </form>
   )
 }
